% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Gets the means of many samples}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{The vector of numeric data to sample}

\item{n}{The size of the sample}

\item{reps}{The number of samples to take}
}
\value{
A vector of double representing the means of each random sample from
the provided data
}
\description{
Gets the means of many samples
}
