% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Gets the mean of a random sample}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{The vector of numeric data to sample}

\item{n}{The size of the sample}
}
\value{
A double representing the mean of a random sample from the provided data
}
\description{
Gets the mean of a random sample
}
