% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Gets the means of many random samples of differing sizes}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{The vector of numeric data to sample}

\item{reps}{The number of samples to take at each size}

\item{ns}{A vector of sizes for each sample to be taken at}
}
\value{
A vector of doubles representing the mean of random sample at each
size from the provided data
}
\description{
Gets the means of many random samples of differing sizes
}
